planning: fix more float / double conversion warning.

+++ modules/planning/toolkits/optimizers/dp_st_speed/dp_st_cost.cc
@@ -29,7 +29,7 @@
 namespace apollo {
 namespace planning {
 namespace {
-constexpr float kInf = std::numeric_limits<float>::infinity();
+constexpr double kInf = std::numeric_limits<double>::infinity();
 }
 
 DpStCost::DpStCost(const DpStSpeedConfig& config,
@@ -63,15 +63,15 @@ void DpStCost::AddToKeepClearRange(
       continue;
     }
 
-    float start_s = obstacle->st_boundary().min_s();
-    float end_s = obstacle->st_boundary().max_s();
+    double start_s = obstacle->st_boundary().min_s();
+    double end_s = obstacle->st_boundary().max_s();
     keep_clear_range_.emplace_back(start_s, end_s);
   }
   SortAndMergeRange(&keep_clear_range_);
 }
 
 void DpStCost::SortAndMergeRange(
-    std::vector<std::pair<float, float>>* keep_clear_range) {
+    std::vector<std::pair<double, double>>* keep_clear_range) {
   if (!keep_clear_range || keep_clear_range->empty()) {
     return;
   }
@@ -91,7 +91,7 @@ void DpStCost::SortAndMergeRange(
   keep_clear_range->resize(i + 1);
 }
 
-bool DpStCost::InKeepClearRange(float s) const {
+bool DpStCost::InKeepClearRange(double s) const {
   if (keep_clear_range_.empty()) {
     return false;
   }
@@ -103,18 +103,18 @@ bool DpStCost::InKeepClearRange(float s) const {
   return false;
 }
 
-float DpStCost::GetObstacleCost(const StGraphPoint& st_graph_point) {
-  const float s = st_graph_point.point().s();
-  const float t = st_graph_point.point().t();
+double DpStCost::GetObstacleCost(const StGraphPoint& st_graph_point) {
+  const double s = st_graph_point.point().s();
+  const double t = st_graph_point.point().t();
 
-  float cost = 0.0;
+  double cost = 0.0;
   for (const auto* obstacle : obstacles_) {
     if (!obstacle->IsBlockingObstacle()) {
       continue;
     }
 
     auto boundary = obstacle->st_boundary();
-    const float kIgnoreDistance = 200.0;
+    const double kIgnoreDistance = 200.0;
     if (boundary.min_s() > kIgnoreDistance) {
       continue;
     }
@@ -138,8 +138,8 @@ float DpStCost::GetObstacleCost(const StGraphPoint& st_graph_point) {
       s_lower = boundary_cost_[boundary_index][st_graph_point.index_t()].second;
     }
     if (s < s_lower) {
-      constexpr float kSafeTimeBuffer = 3.0;
-      const float len = obstacle->speed() * kSafeTimeBuffer;
+      constexpr double kSafeTimeBuffer = 3.0;
+      const double len = obstacle->speed() * kSafeTimeBuffer;
       if (s + len < s_lower) {
         continue;
       } else {
@@ -147,7 +147,7 @@ float DpStCost::GetObstacleCost(const StGraphPoint& st_graph_point) {
                 std::pow((len - s_lower + s), 2);
       }
     } else if (s > s_upper) {
-      const float kSafeDistance = 20.0;  // or calculated from velocity
+      const double kSafeDistance = 20.0;  // or calculated from velocity
       if (s > s_upper + kSafeDistance) {
         continue;
       } else {
@@ -159,16 +159,16 @@ float DpStCost::GetObstacleCost(const StGraphPoint& st_graph_point) {
   return cost * unit_t_;
 }
 
-float DpStCost::GetReferenceCost(const STPoint& point,
+double DpStCost::GetReferenceCost(const STPoint& point,
                                  const STPoint& reference_point) const {
   return config_.reference_weight() * (point.s() - reference_point.s()) *
          (point.s() - reference_point.s()) * unit_t_;
 }
 
-float DpStCost::GetSpeedCost(const STPoint& first, const STPoint& second,
-                             const float speed_limit) const {
-  float cost = 0.0;
-  const float speed = (second.s() - first.s()) / unit_t_;
+double DpStCost::GetSpeedCost(const STPoint& first, const STPoint& second,
+                             const double speed_limit) const {
+  double cost = 0.0;
+  const double speed = (second.s() - first.s()) / unit_t_;
   if (speed < 0) {
     return kInf;
   }
@@ -179,7 +179,7 @@ float DpStCost::GetSpeedCost(const STPoint& first, const STPoint& second,
             config_.default_speed_cost();
   }
 
-  float det_speed = (speed - speed_limit) / speed_limit;
+  double det_speed = (speed - speed_limit) / speed_limit;
   if (det_speed > 0) {
     cost += config_.exceed_speed_penalty() * config_.default_speed_cost() *
             fabs(speed * speed) * unit_t_;
@@ -190,9 +190,9 @@ float DpStCost::GetSpeedCost(const STPoint& first, const STPoint& second,
   return cost;
 }
 
-float DpStCost::GetAccelCost(const float accel) {
-  float cost = 0.0;
-  constexpr float kEpsilon = 0.1;
+double DpStCost::GetAccelCost(const double accel) {
+  double cost = 0.0;
+  constexpr double kEpsilon = 0.1;
   constexpr size_t kShift = 100;
   const size_t accel_key = static_cast<size_t>(accel / kEpsilon + 0.5 + kShift);
   DCHECK_LT(accel_key, accel_cost_.size());
@@ -201,11 +201,11 @@ float DpStCost::GetAccelCost(const float accel) {
   }
 
   if (accel_cost_.at(accel_key) < 0.0) {
-    const float accel_sq = accel * accel;
-    float max_acc = config_.max_acceleration();
-    float max_dec = config_.max_deceleration();
-    float accel_penalty = config_.accel_penalty();
-    float decel_penalty = config_.decel_penalty();
+    const double accel_sq = accel * accel;
+    double max_acc = config_.max_acceleration();
+    double max_dec = config_.max_deceleration();
+    double accel_penalty = config_.accel_penalty();
+    double decel_penalty = config_.decel_penalty();
 
     if (accel > 0.0) {
       cost = accel_penalty * accel_sq;
@@ -223,24 +223,24 @@ float DpStCost::GetAccelCost(const float accel) {
   return cost * unit_t_;
 }
 
-float DpStCost::GetAccelCostByThreePoints(const STPoint& first,
+double DpStCost::GetAccelCostByThreePoints(const STPoint& first,
                                           const STPoint& second,
                                           const STPoint& third) {
-  float accel = (first.s() + third.s() - 2 * second.s()) / (unit_t_ * unit_t_);
+  double accel = (first.s() + third.s() - 2 * second.s()) / (unit_t_ * unit_t_);
   return GetAccelCost(accel);
 }
 
-float DpStCost::GetAccelCostByTwoPoints(const float pre_speed,
+double DpStCost::GetAccelCostByTwoPoints(const double pre_speed,
                                         const STPoint& pre_point,
                                         const STPoint& curr_point) {
-  float current_speed = (curr_point.s() - pre_point.s()) / unit_t_;
-  float accel = (current_speed - pre_speed) / unit_t_;
+  double current_speed = (curr_point.s() - pre_point.s()) / unit_t_;
+  double accel = (current_speed - pre_speed) / unit_t_;
   return GetAccelCost(accel);
 }
 
-float DpStCost::JerkCost(const float jerk) {
-  float cost = 0.0;
-  constexpr float kEpsilon = 0.1;
+double DpStCost::JerkCost(const double jerk) {
+  double cost = 0.0;
+  constexpr double kEpsilon = 0.1;
   constexpr size_t kShift = 200;
   const size_t jerk_key = static_cast<size_t>(jerk / kEpsilon + 0.5 + kShift);
   if (jerk_key >= jerk_cost_.size()) {
@@ -248,7 +248,7 @@ float DpStCost::JerkCost(const float jerk) {
   }
 
   if (jerk_cost_.at(jerk_key) < 0.0) {
-    float jerk_sq = jerk * jerk;
+    double jerk_sq = jerk * jerk;
     if (jerk > 0) {
       cost = config_.positive_jerk_coeff() * jerk_sq * unit_t_;
     } else {
@@ -263,34 +263,34 @@ float DpStCost::JerkCost(const float jerk) {
   return cost;
 }
 
-float DpStCost::GetJerkCostByFourPoints(const STPoint& first,
+double DpStCost::GetJerkCostByFourPoints(const STPoint& first,
                                         const STPoint& second,
                                         const STPoint& third,
                                         const STPoint& fourth) {
-  float jerk = (fourth.s() - 3 * third.s() + 3 * second.s() - first.s()) /
+  double jerk = (fourth.s() - 3 * third.s() + 3 * second.s() - first.s()) /
                (unit_t_ * unit_t_ * unit_t_);
   return JerkCost(jerk);
 }
 
-float DpStCost::GetJerkCostByTwoPoints(const float pre_speed,
-                                       const float pre_acc,
+double DpStCost::GetJerkCostByTwoPoints(const double pre_speed,
+                                       const double pre_acc,
                                        const STPoint& pre_point,
                                        const STPoint& curr_point) {
-  const float curr_speed = (curr_point.s() - pre_point.s()) / unit_t_;
-  const float curr_accel = (curr_speed - pre_speed) / unit_t_;
-  const float jerk = (curr_accel - pre_acc) / unit_t_;
+  const double curr_speed = (curr_point.s() - pre_point.s()) / unit_t_;
+  const double curr_accel = (curr_speed - pre_speed) / unit_t_;
+  const double jerk = (curr_accel - pre_acc) / unit_t_;
   return JerkCost(jerk);
 }
 
-float DpStCost::GetJerkCostByThreePoints(const float first_speed,
+double DpStCost::GetJerkCostByThreePoints(const double first_speed,
                                          const STPoint& first,
                                          const STPoint& second,
                                          const STPoint& third) {
-  const float pre_speed = (second.s() - first.s()) / unit_t_;
-  const float pre_acc = (pre_speed - first_speed) / unit_t_;
-  const float curr_speed = (third.s() - second.s()) / unit_t_;
-  const float curr_acc = (curr_speed - pre_speed) / unit_t_;
-  const float jerk = (curr_acc - pre_acc) / unit_t_;
+  const double pre_speed = (second.s() - first.s()) / unit_t_;
+  const double pre_acc = (pre_speed - first_speed) / unit_t_;
+  const double curr_speed = (third.s() - second.s()) / unit_t_;
+  const double curr_acc = (curr_speed - pre_speed) / unit_t_;
+  const double jerk = (curr_acc - pre_acc) / unit_t_;
   return JerkCost(jerk);
 }
