+++ modules/planning/toolkits/optimizers/dp_st_speed/dp_st_speed_optimizer.cc
@@ -105,7 +105,7 @@ bool DpStSpeedOptimizer::SearchStGraph(
     return false;
   }
 
-  const float path_length = path_data.discretized_path().Length();
+  const double path_length = path_data.discretized_path().Length();
   StGraphData st_graph_data(boundaries, init_point_, speed_limit, path_length);
 
   DpStGraph st_graph(st_graph_data, dp_st_speed_config_,
